SFDM-3243 Доработка процесса data-pipeline, добавление шага публикаци в платформу наблюдаемости

See merge request products/eafdmmart/fdm-bpm!147

+++ src/main/java/ru/beeline/fdmbpm/client/ViewPlatformClient.java
@@ -38,7 +38,7 @@ public class ViewPlatformClient {
         }
     }
 
-    public void postStructurizrModel(String token, Integer modelId) {
+    public void postStructurizrModel(String token, Long modelId) {
         try {
             HttpHeaders headers = new HttpHeaders();
             headers.setContentType(MediaType.APPLICATION_JSON);
